% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgIslPlot_Mval.R
\name{avgIslPlot_Mval}
\alias{avgIslPlot_Mval}
\title{Mvalue Ploting}
\usage{
avgIslPlot_Mval(dat, sqam, IslName, ClockSet, includeBar = TRUE,
  bottomleft = "")
}
\arguments{
\item{dat}{contains a dataframe with illumina 450k methylation data plots non-black cirlced points}

\item{IslName}{is the name of cpg island}

\item{squam}{second dataframe with same type as dat used to plot black circled points}

\item{Clockset}{dataframe including info on CpG clocks on Islands}
}
\description{
This function takes multiple M-values for two datasets, averages the m-value at each cpg for each dataset ,and connects
the two  sets of averages with a line at each cpg location.
}
\examples{
avgIslPlot_Mval(dat = beM.64, sqam = beM.64, IslName = CpGIslName, ClockSet = `CpGs.hypo.drift10.isl`, includeBar = TRUE, bottomleft = "")
}
\keyword{methylation}

