% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IslandPlotMval.R
\name{IslandPlot_Mval}
\alias{IslandPlot_Mval}
\title{Mvalue Ploting}
\usage{
IslandPlot_Mval(dat, IslName, ClockSet, bottomleft = "")
}
\arguments{
\item{dat}{contains a dataframe with illumina 450k methylation data}

\item{IslName}{name of CpG island you want to graph}

\item{Clockset}{dataframe including info on CpG clocks on Islands}
}
\description{
Plot multiple CpG M-values for an island,shore,and shelf.
}
\examples{
IslandPlot_Mval(beM.64, "chr7:77648451-77648668",  "CpGs.hypo.drift10.isl")
}
\keyword{M_val}
\keyword{islands}

